refactor: unify event system and restructure blocking/platform modules
Replace separate FocusChanged + ActivityChanged D-Bus signals with a
single Event signal carrying a (uussuu) struct with typed tags
(Focus, Unfocus, Block, Idle, Resume, LogOut, PowerEvent).  Every
PlatformEvent is now UID-scoped; system-wide events are gone.

Blocking module restructuring:
  - buffer.rs -> core/buffer.rs (actor-owned event buffer)
  - domain.rs  -> domain/{mod, timed_event}.rs
  - data/persistence/deltas/ -> data/deltas/ (flattened hierarchy)
  - Added data/repo.rs with BlockingRepo
  - Removed data/persistence/events.rs, interval_split.rs, pre_buffer.rs

Platform layer cleanup:
  - Removed screen_lock.rs and suspend.rs Linux modules
  - Power/screen-lock events flow through unified PlatformEvent::PowerEvent
  - PluginRegistry owns all plugin lifecycle

Daemon main wiring:
  - Removed separate PowerStateWatcher / ScreenLockWatcher spawns
  - EnforcerActor receives PluginRegistry directly

Hyprland plugin:
  - Unified Event signal replaces FocusChanged + ActivityChanged
  - New EventTag enum and (uussuu) struct serialization

D-Bus constants and domain tests updated for unified schema.

+++ crates/daemon/src/blocking/data/queries.rs
@@ -7,7 +7,7 @@ use wellbeing_core::{AppId, CategoryId, Clock, Uid};
 use crate::policy::{DieselPolicyRepo, Policy, PolicyRepo as _};
 use crate::store::schema;
 
-use super::events::BlockingRepo;
+use super::repo::BlockingRepo;
 
 impl BlockingRepo {
     /// Resolve category ids for an app (user-specific then fallback).
@@ -46,7 +46,7 @@ impl BlockingRepo {
     ) -> anyhow::Result<i64> {
         let mut conn = self.pool.get().await?;
         let today = clock.now().format("%Y-%m-%d").to_string();
-        let row: Option<(i32, i32)> = schema::daily_usage::table
+        let row: Option<(i64, i64)> = schema::daily_usage::table
             .filter(schema::daily_usage::date.eq(&today))
             .filter(schema::daily_usage::user_id.eq(uid.0 as i32))
             .filter(schema::daily_usage::app_id.eq(app_id.as_ref()))
@@ -59,6 +59,6 @@ impl BlockingRepo {
             .ok();
 
         let (closed, open) = row.unwrap_or((0, 0));
-        Ok(closed as i64 + open as i64)
+        Ok(closed + open)
     }
 }
